[Wasm64] Add support for threads (#18705)
Testing wasm64 with threads requires a nightly build of node (since
it depends on some very recent fixes on the v8 side).

+++ test/core/pthread/exceptions.cpp
@@ -43,7 +43,7 @@ pthread_t thread[NUM_THREADS];
 
 void CreateThread(int i)
 {
-  int rc = pthread_create(&thread[i], nullptr, ThreadMain, (void*)i);
+  int rc = pthread_create(&thread[i], nullptr, ThreadMain, (void*)(intptr_t)i);
   assert(rc == 0);
 }
